feature/changePassword: extract components from UserDataPage into UserDataTabs & PasswordTabs, better error handling, improve text for InfoPage
